% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoi-package.R
\docType{package}
\name{qoi-package}
\alias{qoi}
\alias{qoi-package}
\title{qoi: Read and Write QOI Images}
\description{
The new QOI file format offers a very simple but efficient image compression algorithm. This package provides an easy and simple way to read, write and display bitmap images stored in the QOI (Quite Ok Image) format. It can read and write both files and in-memory raw vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JohannesFriedrich/qoi4R}
  \item Report bugs at \url{https://github.com/JohannesFriedrich/qoi4R/issues}
}

}
\author{
\strong{Maintainer}: Johannes Friedrich \email{Johannes.Friedrich@posteo.de} [translator]

Other contributors:
\itemize{
  \item Dominic Szablewski (C library 'qoi') [copyright holder]
}

}
\keyword{internal}
