% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qoi-package.R
\docType{data}
\name{Rlogo_RGBA}
\alias{Rlogo_RGBA}
\title{RGBA values for the Rlogo (https://www.r-project.org/logo/)}
\format{
[matrix] with 561 x 724 x 4 elements
}
\description{
RGBA values for the Rlogo (https://www.r-project.org/logo/)
}
\author{
Johannes Friedrich
}
